Refactored Wiremock setup.
