+++ server/eclipse-project/src/test/resources/test.properties
@@ -22,5 +22,4 @@ corla_https_port = 8889
 
 # port numbers for mocked raire services, used for testing endpoints. These need to be all different,
 # from each other and from anything else running on the computer.
-get_assertions_mock_port = 8111
-generate_assertions_mock_port = 8110
+raire_mock_port = 8112
